+++ session/session.go
@@ -14,7 +14,6 @@ import (
 	"os"
 	"os/exec"
 	"reflect"
-	"sort"
 	"strings"
 
 	"github.com/jedib0t/go-pretty/v6/table"
@@ -96,11 +95,6 @@ type Session struct {
 	incltoken bool
 	treefile  string
 	evalfile  string
-
-	//historyExpr		[]ast.Expression
-	//historyStmsts		[]ast.Statement
-	//historyPrograms	[]ast.Programs
-	// --> maybe not needed, maybe we should put the Stmts programs consist of into historyStmts
 }
 
 const ( //default settings
@@ -141,253 +135,6 @@ func NewSession(in io.Reader, out io.Writer) *Session {
 	return s
 }
 
-type command struct {
-	name     string
-	single   func()
-	with_arg func(string) // initialized here --> end msg about potential cycle
-	usage    []struct {
-		args string
-		msg  string
-	}
-}
-
-var commands = make(map[string]command)
-
-func (s *Session) init() { // to avoid cycle
-
-	c_quit := &command{
-		name:   "q[uit]",
-		single: s.exec_quit,
-		usage: []struct {
-			args string
-			msg  string
-		}{
-			{"~", "quit the session"},
-		},
-	}
-	commands["quit"] = *c_quit
-	commands["q"] = commands["quit"]
-
-	c_settings := &command{
-		name:   "settings",
-		single: s.exec_settings,
-		usage: []struct {
-			args string
-			msg  string
-		}{
-			{"~", "list all settings with their current values and defaults"},
-		},
-	}
-	commands["settings"] = *c_settings
-
-	c_clear := &command{
-		name:   "c[lear]",
-		single: s.exec_clear,
-		usage: []struct {
-			args string
-			msg  string
-		}{
-			{"~", "clear the environment"},
-		},
-	}
-	commands["clear"] = *c_clear
-	commands["c"] = commands["clear"]
-
-	c_clearscreen := &command{
-		name:   "cl[earscreen]",
-		single: s.exec_clearscreen,
-		usage: []struct {
-			args string
-			msg  string
-		}{
-			{"~", "clear the terminal screen"},
-		},
-	}
-	commands["clearscreen"] = *c_clearscreen
-	commands["cl"] = commands["clearscreen"]
-
-	c_list := &command{
-		name:   "l[ist]",
-		single: s.exec_list,
-		usage: []struct {
-			args string
-			msg  string
-		}{
-			{"~", "list all identifiers in the environment alphabetically \n\t with types and values"},
-		},
-	}
-	commands["list"] = *c_list
-	commands["l"] = commands["list"]
-
-	c_paste := &command{
-		name:     "paste",
-		with_arg: s.exec_paste,
-		single:   s.exec_paste_empty_arg,
-		usage: []struct {
-			args string
-			msg  string
-		}{
-			{"~ <input>", "evaluate multiline <input> (terminated by blank line)"},
-		},
-	}
-	commands["paste"] = *c_paste
-
-	c_set := &command{
-		name:     "set",
-		with_arg: s.exec_set,
-		usage: []struct {
-			args string
-			msg  string
-		}{
-			{"~ process <p>", "<p> must be: [e]val, [p]arse, [t]ype"},
-			{"~ level <l>", "<l> must be: [p]rogram, [s]tatement, [e]xpression"},
-			{"~ logparse", "additionally output ast-string"},
-			{"~ logtype", "additionally output objecttype"},
-			{"~ logtrace", "additionally output evaluation trace"},
-			{"~ incltoken", "include tokens in representations of asts"},
-			{"~ paste", "enable multiline support"},
-			{"~ prompt <prompt>", "set prompt string to <prompt>"},
-			{"~ treefile <f>", "set file that outputs ast-pdfs to <f>"},
-			{"~ evalfile <f>", "set file that outputs eval-pdfs to <f>"},
-		},
-	}
-	commands["set"] = *c_set
-
-	c_reset := &command{
-		name:     "reset",
-		with_arg: s.exec_reset,
-		usage: []struct {
-			args string
-			msg  string
-		}{
-			{"~ <setting>", "set <setting> to default\n\t for an overview consult :settings and/or :h set"},
-		},
-	}
-	commands["reset"] = *c_reset
-
-	c_unset := &command{
-		name:     "unset",
-		with_arg: s.exec_unset,
-		usage: []struct {
-			args string
-			msg  string
-		}{
-			{"~ <setting>", "set boolean <setting> to false\n\t for an overview consult :settings and/or :h set"},
-			//	{"~ logparse", "don't additionally output ast-string"},
-			//	{"~ logtype", "don't additionally output objecttype"},
-			//	{"~ paste", "disable multiline support"},
-			//incltoken logtrace
-		},
-	}
-	commands["unset"] = *c_unset
-
-	c_help := &command{
-		name:     "h[elp]",
-		single:   s.exec_help_all,
-		with_arg: s.exec_help,
-		usage: []struct {
-			args string
-			msg  string
-		}{
-			{"~", "list all commands with usage"},
-			{"~ <cmd>", "print usage command <cmd>"},
-		},
-	}
-	commands["help"] = *c_help
-	commands["h"] = commands["help"]
-
-	c_type := &command{
-		name:     "t[ype]",
-		with_arg: s.exec_type,
-		usage: []struct {
-			args string
-			msg  string
-		}{
-			{"~ <input>", "show objecttype <input> evaluates to"},
-		},
-	}
-	commands["type"] = *c_type
-	commands["t"] = commands["type"]
-
-	c_trace := &command{
-		name:     "trace",
-		with_arg: s.exec_trace,
-		usage: []struct {
-			args string
-			msg  string
-		}{
-			{"~ <input>", "show evaluation trace step by step"},
-		},
-	}
-	commands["trace"] = *c_trace
-
-	c_parse := &command{
-		name:     "p[arse]",
-		with_arg: s.exec_parse,
-		usage: []struct {
-			args string
-			msg  string
-		}{
-			{"~ <input>", "parse <input>"},
-		},
-	}
-	commands["parse"] = *c_parse
-	commands["p"] = commands["parse"]
-
-	c_eval := &command{
-		name:     "e[val]",
-		with_arg: s.exec_eval,
-		usage: []struct {
-			args string
-			msg  string
-		}{
-			{"~ <input>", "print out value of object <input> evaluates to"},
-		},
-	}
-	commands["eval"] = *c_eval
-	commands["e"] = commands["eval"]
-
-	c_expr := &command{
-		name:     "expr[ession]",
-		with_arg: s.exec_expression,
-		usage: []struct {
-			args string
-			msg  string
-		}{
-			{"~ <input>", "expect <input> to be an expression"},
-		},
-	}
-	commands["expression"] = *c_expr
-	commands["expr"] = commands["expression"]
-
-	c_stmt := &command{
-		name:     "stmt|statement",
-		with_arg: s.exec_statement,
-		usage: []struct {
-			args string
-			msg  string
-		}{
-			{"~ <input>", "expect <input> to be a statement"},
-		},
-	}
-	commands["statement"] = *c_stmt
-	commands["stmt"] = commands["statement"]
-
-	c_prog := &command{
-		name:     "prog[ram]",
-		with_arg: s.exec_program,
-		usage: []struct {
-			args string
-			msg  string
-		}{
-			{"~ <input>", "expect <input> to be a program"},
-		},
-	}
-	commands["program"] = *c_prog
-	commands["prog"] = commands["program"]
-
-}
-
 // decide which function
 func (s *Session) exec_cmd(line string) {
 	if !strings.HasPrefix(line, ":") {
@@ -398,18 +145,16 @@ func (s *Session) exec_cmd(line string) {
 	slice := strings.SplitN(line, " ", 2)
 
 	cmd := slice[0]
-	if c_entry, ok := commands[cmd]; ok {
-		if len(slice) == 1 {
-			if c_entry.single != nil {
-				c_entry.single()
-				return
-			}
-		} else {
-			arg := slice[1]
-			if c_entry.with_arg != nil {
-				c_entry.with_arg(arg)
-				return
-			}
+	if len(slice) == 1 {
+		if exec, ok := commands.get_exec_single(cmd); ok {
+			exec()
+			return
+		}
+	} else {
+		arg := slice[1]
+		if exec, ok := commands.get_exec_with_arg(cmd); ok {
+			exec(arg)
+			return
 		}
 	}
 
@@ -456,27 +201,8 @@ func (s *Session) exec_list() {
 // commands
 func (s *Session) exec_help(cmd string) {
 
-	if command, ok := commands[cmd]; ok {
-
-		//print
-		t := table.NewWriter()
-		t.SetOutputMirror(s.out)
-
-		usage := command.usage
-		if len(usage) == 0 {
-			t.AppendRow([]interface{}{command.name, "no usage message provided"})
-		} else {
-			for i, msg := range usage {
-				if i == 0 {
-					t.AppendRow([]interface{}{command.name, msg.args, msg.msg})
-				} else {
-					t.AppendRow([]interface{}{"", msg.args, msg.msg})
-				}
-			}
-
-		}
-		t.Render()
-
+	if usage, ok := commands.usage(cmd); ok {
+		fmt.Fprint(s.out, usage)
 		return
 	}
 
@@ -484,48 +210,7 @@ func (s *Session) exec_help(cmd string) {
 }
 
 func (s *Session) exec_help_all() {
-
-	// print out all commands, but only once (--> quit, q)
-	// always in the same order
-
-	//remove duplicates
-	var set = make(map[string]command)
-	for _, command := range commands {
-		set[command.name] = command
-	}
-
-	//sort
-	names := make([]string, len(set))
-	i := 0
-	for name := range set {
-		names[i] = name
-		i++
-	}
-	sort.Strings(names)
-
-	//print
-	t := table.NewWriter()
-	t.SetOutputMirror(s.out)
-	t.AppendHeader(table.Row{"Name", "", "Usage"})
-	t.AppendSeparator()
-
-	for _, name := range names {
-		command := set[name]
-		usage := command.usage
-		if len(usage) == 0 {
-			t.AppendRow([]interface{}{name, "no usage message provided"})
-		} else {
-			for i, msg := range usage {
-				if i == 0 {
-					t.AppendRow([]interface{}{name, msg.args, msg.msg})
-				} else {
-					t.AppendRow([]interface{}{"", msg.args, msg.msg})
-				}
-			}
-
-		}
-	}
-	t.Render()
+	fmt.Fprint(s.out, commands.menu())
 }
 
 // settings