+++ changelog.md
@@ -10,11 +10,11 @@ Started late: on 2021-04-20
   - add changelog
   - restructure main README
 
-### 2021-04-22
+### 2021-04-23
 
 - reorganize instruction set
   - add c[lear]
-
+  - choose order of commands in menu
 
 ## [Summary of what happened before 2021-04-20]